LayoutTests/http/tests/iframe-monitor/workers/shared-worker.html

<!DOCTYPE html> <!-- webkit-test-runner [ IFrameResourceMonitoringEnabled=true ] -->
<html>

<head>
    <script src="/js-test-resources/js-test.js"></script>
    <script src="../resources/monitor-setup.js"></script>
</head>

<body>
    <script>

        description("Test iframes using same shared worker are unloaded.");
        window.jsTestIsAsync = true;

        onload = afterSetup(async () => {
            // Make sure iframe load is done after rule is set correctly.
            const stage = document.querySelector('#stage');
            const base = 'http://localhost:8080/iframe-monitor/resources';

            let ready = 0;
            window.addEventListener('message', (e) => {
                ready++;
            });

            stage.innerHTML = `
        <iframe name="frame1" src="${base}/--eligible--/shared-worker.html"></iframe>
        <iframe name="frame2" src="${base}/--eligible--/shared-worker.html"></iframe>
    `;

            while (ready < 2)
                await pause(10);

            // Send message to one of iframe to start fetching via shared worker.
            const frame1 = document.querySelector('iframe[name=frame1]');
            frame1.contentWindow.postMessage(20 * 1024, '*');

            await waitUntilUnload('frame1');
            await waitUntilUnload('frame2');

            shouldNotBe(`document.querySelector('iframe[name=frame1]').srcdoc`, '""');
            shouldNotBe(`document.querySelector('iframe[name=frame2]').srcdoc`, '""');

            finishJSTest();
        });

    </script>

    <div id="stage"></div>
</body>

</html>
